Add calculator functionality
Add click event listener to button housing
instead of indiviual buttons to improve
performance by using event delegation.

Add generalised classes to each button in
button housing to easily target buttons with
event delegation.

Add evaluateExpression function to convert
textContent from upper-display class to array
split by whitespace. Use while loop and chain if
statements to calculate expression while
following BODMAS/PEMDAS and sign rules.

This is done by finding the index of the
arithmetic operator with the highest priority in
BODMAS/PEMDAS rule, getting the numbers in the
index before and after the operator in the
array, passing the values to operate function to
calculate the answer to the expression and
replacing the two numbers and operator in the
array with answer returned by operate function
using array splice.

This is repeated using while loop until the
array length is one which represents the final
answer.

If statements for Add and Substract operators in
evaluateExpression function have to account for
sign rule by checking if there is a negative sign
2 index spaces to the left of an Add or Substract
operator and converting the first number for the
operate function to negative and converting the
answer returned to positive when splicing the
answer into the array.

The answers are converted to positive to keep
compatilibity with the multiply and divide
operators in the evaluateExpression function.